+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Added
 - Support to comments (#163)
 
+## Fixed
+- Update options without color (#166)
+
 ## Documentation
 - Document `People` page property (#162)